Move resolve_path to rustc_builtin_macros and make it private
